RF: Move CreateDialog to sync
